requirements.txt
asgiref==3.2.5
auditlog3==1.0.1
boto3==1.14.56
botocore==1.17.56
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.4
django-appconf==1.0.4
django-cors-headers==3.2.1
django-extensions==2.2.9
django-ipware==2.1.0
django-model-utils==4.0.0
django-reset-migrations==0.4.0
django-rest-swagger==2.2.0
django-storages==1.10
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
docutils==0.15.2
idna==2.8
imagekitio==2.2.3
itypes==1.1.0
Jinja2==2.11.1
jmespath==0.10.0
jsonfield==3.1.0
MarkupSafe==1.1.1
mysqlclient==1.4.6
openapi-codec==1.3.2
Pillow==7.2.0
PyJWT==1.7.1
python-dateutil==2.8.1
python-http-client==3.2.7
pytz==2019.3
requests==2.22.0
s3transfer==0.3.3
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.1
stripe==2.55.0
uritemplate==3.0.1
urllib3==1.25.8

